Create BinaryToDecimal
A recursive function that takes a string representation of an unsigned binary integer and returns a decimal integer representation of the binary string

This is a recreation of code made for a Computer Science II Exam written in <5 minutes. I felt this was similar to something I have seen in job interviews, so I figured this would be a fun piece of code to add to my GitHub. Nothing remotely difficult, but I figured it shows that I understand the basics and can quickly solve novel problems with intuitive solutions.